+++ feathers/utils/__init__.py
@@ -1,0 +1,3 @@
+from ._errors import friendly_list
+
+__all__ = ["friendly_list"]