+++ feathers/widgets/__init__.py
@@ -1,4 +1,5 @@
-from ._chat import Chat, ChatEntry, ChatStyle, Participant
+from ._cached_view import CachedView
+from ._nav_view import NavigableView
 from .help import Help, HelpEntry, HelpProvider
 
-__all__ = ["Help", "HelpEntry", "HelpProvider", "Chat", "ChatEntry", "ChatStyle", "Participant"]
+__all__ = ["Help", "HelpEntry", "HelpProvider", "NavigableView", "CachedView"]